Endre brevkode til å følge confluence struktur
