+++ apps/etterlatte-brev-api/src/main/kotlin/no/nav/etterlatte/brev/brevbaker/BrevbakerRequest.kt
@@ -53,9 +53,8 @@ enum class EtterlatteBrevKode {
     BARNEPENSJON_AVSLAG_IKKEYRKESSKADE,
     BARNEPENSJON_INNVILGELSE,
     BARNEPENSJON_INNVILGELSE_NY,
-    OMS_INNVILGELSE_FOERSTEGANGSVEDTAK,
-    OMS_INNVILGELSE_FOERSTEGANGSVEDTAK_UTFALL,
-    OMS_INNVILGELSE_AUTO,
+    OMS_FOERSTEGANGSVEDTAK_INNVILGELSE,
+    OMS_FOERSTEGANGSVEDTAK_INNVILGELSE_UTFALL,
     OMS_INNVILGELSE_MANUELL,
     OMS_OPPHOER_MANUELL,
     BARNEPENSJON_REVURDERING_ADOPSJON,
