+++ apps/etterlatte-brev-api/src/main/kotlin/no/nav/etterlatte/brev/model/BrevDataMapper.kt
@@ -15,8 +15,8 @@ import no.nav.etterlatte.brev.brevbaker.EtterlatteBrevKode.BARNEPENSJON_REVURDER
 import no.nav.etterlatte.brev.brevbaker.EtterlatteBrevKode.BARNEPENSJON_REVURDERING_SOESKENJUSTERING
 import no.nav.etterlatte.brev.brevbaker.EtterlatteBrevKode.BARNEPENSJON_REVURDERING_UT_AV_FENGSEL
 import no.nav.etterlatte.brev.brevbaker.EtterlatteBrevKode.BARNEPENSJON_REVURDERING_YRKESSKADE
-import no.nav.etterlatte.brev.brevbaker.EtterlatteBrevKode.OMS_INNVILGELSE_FOERSTEGANGSVEDTAK
-import no.nav.etterlatte.brev.brevbaker.EtterlatteBrevKode.OMS_INNVILGELSE_FOERSTEGANGSVEDTAK_UTFALL
+import no.nav.etterlatte.brev.brevbaker.EtterlatteBrevKode.OMS_FOERSTEGANGSVEDTAK_INNVILGELSE
+import no.nav.etterlatte.brev.brevbaker.EtterlatteBrevKode.OMS_FOERSTEGANGSVEDTAK_INNVILGELSE_UTFALL
 import no.nav.etterlatte.brev.brevbaker.EtterlatteBrevKode.OMS_OPPHOER_MANUELL
 import no.nav.etterlatte.funksjonsbrytere.FeatureToggle
 import no.nav.etterlatte.funksjonsbrytere.FeatureToggleService
@@ -85,8 +85,8 @@ class BrevDataMapper(private val featureToggleService: FeatureToggleService) {
         SakType.OMSTILLINGSSTOENAD -> {
             when (val vedtakType = behandling.vedtak.type) {
                 VedtakType.INNVILGELSE -> BrevkodePar(
-                    OMS_INNVILGELSE_FOERSTEGANGSVEDTAK_UTFALL,
-                    OMS_INNVILGELSE_FOERSTEGANGSVEDTAK
+                    OMS_FOERSTEGANGSVEDTAK_INNVILGELSE_UTFALL,
+                    OMS_FOERSTEGANGSVEDTAK_INNVILGELSE
                 )
                 VedtakType.AVSLAG -> TODO("Vedtakstype er ikke støttet: $vedtakType")
                 VedtakType.ENDRING -> TODO("Vedtakstype er ikke støttet: $vedtakType")
@@ -142,7 +142,7 @@ class BrevDataMapper(private val featureToggleService: FeatureToggleService) {
     fun brevDataFerdigstilling(behandling: Behandling, innhold: () -> List<Slate.Element>, kode: BrevkodePar) =
         when (kode.ferdigstilling) {
             BARNEPENSJON_REVURDERING_ENDRING -> EndringHovedmalBrevData.fra(behandling, innhold())
-            OMS_INNVILGELSE_FOERSTEGANGSVEDTAK -> InnvilgetBrevDataOMS.fra(behandling, innhold())
+            OMS_FOERSTEGANGSVEDTAK_INNVILGELSE -> InnvilgetBrevDataOMS.fra(behandling, innhold())
             else ->
                 when (behandling.revurderingsaarsak?.redigerbartBrev) {
                     true -> ManueltBrevData(innhold())
