+++ apps/etterlatte-brev-api/src/test/kotlin/no/nav/etterlatte/brev/VedtaksbrevServiceTest.kt
@@ -145,7 +145,6 @@ internal class VedtaksbrevServiceTest {
         @ParameterizedTest
         @CsvSource(
             value = [
-                "OMSTILLINGSSTOENAD,INNVILGELSE,AUTOMATISK",
                 "OMSTILLINGSSTOENAD,OPPHOER,MANUELL",
                 "OMSTILLINGSSTOENAD,AVSLAG,MANUELL",
                 "OMSTILLINGSSTOENAD,ENDRING,MANUELL",
